fix(tb-334): split dialog file under 1600 lines + fix test mock
- Extract 7 sub-components into CorrectionProposalDialogPanels.tsx (883 lines)
- Main CorrectionProposalDialog.tsx now 1367 lines (was 2212)
- Add reevaluateWithPendingRules mock to ReviewStep.test.tsx
